Refactoring shell to promote scalability and simplicity
