+++ core/src/main/java/tachyon/command/commands/CatCommand.java
@@ -1,0 +1,60 @@
+package tachyon.command.commands;
+
+import java.io.IOException;
+
+import tachyon.TachyonURI;
+import tachyon.client.InStream;
+import tachyon.client.ReadType;
+import tachyon.client.TachyonFS;
+import tachyon.client.TachyonFile;
+import tachyon.command.AbstractCommands;
+
+/**
+ * Prints the file's contents to the console.
+ *
+ */
+public class CatCommand extends AbstractCommands {
+  /**
+   * Prints the file's contents to the console.
+   *
+   * @param argv [] Array of arguments given by the user's input from the terminal
+   * @return 0 if command is successful, -1 if an error occurred.
+   * @throws java.io.IOException
+   */
+  public int cat(String[] argv) throws IOException {
+    if (argv.length != 2) {
+      System.out.println("Usage: tfs cat <path>");
+      return -1;
+    }
+    TachyonURI path = new TachyonURI(argv[1]);
+    TachyonFS tachyonClient = createFS(path);
+    TachyonFile tFile = tachyonClient.getFile(path);
+
+    if (tFile == null) {
+      System.out.println(path + " does not exist.");
+      return -1;
+    }
+    if (tFile.isFile()) {
+      InStream is = tFile.getInStream(ReadType.NO_CACHE);
+      byte[] buf = new byte[512];
+      try {
+        int read = is.read(buf);
+        while (read != -1) {
+          System.out.write(buf, 0, read);
+          read = is.read(buf);
+        }
+      } finally {
+        is.close();
+      }
+      return 0;
+    } else {
+      System.out.println(path + " is not a file.");
+      return -1;
+    }
+  }
+
+  @Override
+  public int execute(String[] argv)  throws IOException {
+    return cat(argv);
+  }
+}
