+++ core/src/main/java/tachyon/command/commands/UnpinCommand.java
@@ -1,0 +1,46 @@
+package tachyon.command.commands;
+
+import java.io.IOException;
+
+import tachyon.TachyonURI;
+import tachyon.client.TachyonFS;
+import tachyon.command.AbstractCommands;
+
+/**
+ * Unpins the given file or folder (recursively unpinning all children if a folder). Pinned files
+ * are never evicted from memory, so this method will allow such files to be evicted.
+ *
+ */
+public class UnpinCommand extends AbstractCommands {
+  @Override
+  public int execute(String[] argv) throws IOException {
+    return unpin(argv);
+  }
+  
+  /**
+   * Unpins the given file or folder (recursively unpinning all children if a folder). Pinned files
+   * are never evicted from memory, so this method will allow such files to be evicted.
+   *
+   * @param argv [] Array of arguments given by the user's input from the terminal
+   * @return 0 if command is successful, -1 if an error occurred.
+   * @throws java.io.IOException
+   */
+  public int unpin(String[] argv) throws IOException {
+    if (argv.length != 2) {
+      System.out.println("Usage: tfs unpin <path>");
+      return -1;
+    }
+    TachyonURI path = new TachyonURI(argv[1]);
+    TachyonFS tachyonClient = createFS(path);
+    int fileId = tachyonClient.getFileId(path);
+    try {
+      tachyonClient.unpinFile(fileId);
+      System.out.println("File '" + path + "' was successfully unpinned.");
+      return 0;
+    } catch (Exception e) {
+      e.printStackTrace();
+      System.out.println("File '" + path + "' could not be unpinned.");
+      return -1;
+    }
+  }
+}
