+++ core/src/test/java/tachyon/command/TFsShellTest.java
@@ -68,8 +68,8 @@ public class TFsShellTest {
   @Test
   public void catDirectoryTest() throws IOException {
     String[] command = new String[] {"mkdir", "/testDir"};
-    mFsShell.mkdir(command);
-    int ret = mFsShell.cat(new String[] {"cat", "/testDir"});
+    mFsShell.run(command);
+    int ret = mFsShell.run(new String[] {"cat", "/testDir"});
     Assert.assertEquals(-1, ret);
     String expected = getCommandOutput(command);
     expected += "/testDir is not a file.\n";
@@ -78,14 +78,14 @@ public class TFsShellTest {
 
   @Test
   public void catNotExistTest() throws IOException {
-    int ret = mFsShell.cat(new String[] {"cat", "/testFile"});
+    int ret = mFsShell.run(new String[] {"cat", "/testFile"});
     Assert.assertEquals(-1, ret);
   }
 
   @Test
   public void catTest() throws IOException {
     TestUtils.createByteFile(mTfs, "/testFile", WriteType.MUST_CACHE, 10);
-    mFsShell.cat(new String[] {"cat", "/testFile"});
+    mFsShell.run(new String[] {"cat", "/testFile"});
     byte expect[] = TestUtils.getIncreasingByteArray(10);
     Assert.assertArrayEquals(expect, mOutput.toByteArray());
   }
@@ -98,7 +98,7 @@ public class TFsShellTest {
     byte toWrite[] = TestUtils.getIncreasingByteArray(SIZE_BYTES);
     fos.write(toWrite);
     fos.close();
-    mFsShell.copyFromLocal(new String[] {"copyFromLocal", testFile.getAbsolutePath(), "/testFile"});
+    mFsShell.run(new String[] {"copyFromLocal", testFile.getAbsolutePath(), "/testFile"});
     Assert.assertEquals(getCommandOutput(new String[] {"copyFromLocal", testFile.getAbsolutePath(),
         "/testFile"}), mOutput.toString());
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/testFile"));
@@ -118,7 +118,7 @@ public class TFsShellTest {
     testDirInner.mkdir();
     File testFile = generateFileContent("/testDir/testFile", TestUtils.getIncreasingByteArray(10));
     generateFileContent("/testDir/testDirInner/testFile2", TestUtils.getIncreasingByteArray(10, 20));
-    mFsShell.copyFromLocal(new String[] {"copyFromLocal", testFile.getParent(), "/testDir"});
+    mFsShell.run(new String[] {"copyFromLocal", testFile.getParent(), "/testDir"});
     Assert.assertEquals(getCommandOutput(new String[] {"copyFromLocal", testFile.getParent(),
         "/testDir"}), mOutput.toString());
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/testDir/testFile"));
@@ -140,7 +140,7 @@ public class TFsShellTest {
         "tachyon://" + mLocalTachyonCluster.getMasterHostname() + ":"
             + mLocalTachyonCluster.getMasterPort() + "/destFileURI";
     // when
-    mFsShell.copyFromLocal(new String[] {"copyFromLocal", testFile.getPath(), tachyonURI});
+    mFsShell.run(new String[] {"copyFromLocal", testFile.getPath(), tachyonURI});
     String cmdOut =
         getCommandOutput(new String[] {"copyFromLocal", testFile.getPath(), tachyonURI});
     // then
@@ -154,7 +154,7 @@ public class TFsShellTest {
   @Test
   public void copyToLocalLargeTest() throws IOException {
     TestUtils.createByteFile(mTfs, "/testFile", WriteType.MUST_CACHE, SIZE_BYTES);
-    mFsShell.copyToLocal(new String[] {"copyToLocal", "/testFile",
+    mFsShell.run(new String[] {"copyToLocal", "/testFile",
         mLocalTachyonCluster.getTachyonHome() + "/testFile"});
     Assert.assertEquals(getCommandOutput(new String[] {"copyToLocal", "/testFile",
         mLocalTachyonCluster.getTachyonHome() + "/testFile"}), mOutput.toString());
@@ -169,7 +169,7 @@ public class TFsShellTest {
   @Test
   public void copyToLocalTest() throws IOException {
     TestUtils.createByteFile(mTfs, "/testFile", WriteType.MUST_CACHE, 10);
-    mFsShell.copyToLocal(new String[] {"copyToLocal", "/testFile",
+    mFsShell.run(new String[] {"copyToLocal", "/testFile",
         mLocalTachyonCluster.getTachyonHome() + "/testFile"});
     Assert.assertEquals(getCommandOutput(new String[] {"copyToLocal", "/testFile",
         mLocalTachyonCluster.getTachyonHome() + "/testFile"}), mOutput.toString());
@@ -186,7 +186,7 @@ public class TFsShellTest {
     TestUtils.createByteFile(mTfs, "/testRoot/testFileA", WriteType.MUST_CACHE, 10);
     TestUtils.createByteFile(mTfs, "/testRoot/testDir/testFileB", WriteType.MUST_CACHE, 20);
     TestUtils.createByteFile(mTfs, "/testRoot/testFileB", WriteType.MUST_CACHE, 30);
-    mFsShell.count(new String[] {"count", "/testRoot"});
+    mFsShell.run(new String[] {"count", "/testRoot"});
     String expected = "";
     String format = "%-25s%-25s%-15s\n";
     expected += String.format(format, "File Count", "Folder Count", "Total Bytes");
@@ -197,7 +197,7 @@ public class TFsShellTest {
   @Test
   public void fileinfoTest() throws IOException {
     int fileId = TestUtils.createByteFile(mTfs, "/testFile", WriteType.MUST_CACHE, 10);
-    mFsShell.fileinfo(new String[] {"fileinfo", "/testFile"});
+    mFsShell.run(new String[] {"fileinfo", "/testFile"});
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/testFile"));
     Assert.assertNotNull(tFile);
     List<ClientBlockInfo> blocks = mTfs.getFileBlocks(fileId);
@@ -267,7 +267,7 @@ public class TFsShellTest {
   @Test
   public void locationTest() throws IOException {
     int fileId = TestUtils.createByteFile(mTfs, "/testFile", WriteType.MUST_CACHE, 10);
-    mFsShell.location(new String[] {"location", "/testFile"});
+    mFsShell.run(new String[] {"location", "/testFile"});
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/testFile"));
     Assert.assertNotNull(tFile);
     List<String> locationsList = tFile.getLocationHosts();
@@ -293,7 +293,7 @@ public class TFsShellTest {
     files[2] = mTfs.getFile(new TachyonURI("/testRoot/testDir/testFileB"));
     int fileIdC = TestUtils.createByteFile(mTfs, "/testRoot/testFileC", WriteType.THROUGH, 30);
     files[3] = mTfs.getFile(fileIdC);
-    mFsShell.ls(new String[] {"count", "/testRoot"});
+    mFsShell.run(new String[] {"ls", "/testRoot"});
     String expected = "";
     String format = "%-10s%-25s%-15s%-5s\n";
     expected +=
@@ -319,7 +319,7 @@ public class TFsShellTest {
     files[1] = mTfs.getFile(new TachyonURI("/testRoot/testDir"));
     int fileIdC = TestUtils.createByteFile(mTfs, "/testRoot/testFileC", WriteType.THROUGH, 30);
     files[2] = mTfs.getFile(fileIdC);
-    mFsShell.ls(new String[] {"count", "/testRoot"});
+    mFsShell.run(new String[] {"ls", "/testRoot"});
     String expected = "";
     String format = "%-10s%-25s%-15s%-5s\n";
     expected +=
@@ -338,7 +338,7 @@ public class TFsShellTest {
 
   @Test
   public void mkdirComplexPathTest() throws IOException {
-    mFsShell.mkdir(new String[] {"mkdir", "/Complex!@#$%^&*()-_=+[]{};\"'<>,.?/File"});
+    mFsShell.run(new String[] {"mkdir", "/Complex!@#$%^&*()-_=+[]{};\"'<>,.?/File"});
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/Complex!@#$%^&*()-_=+[]{};\"'<>,.?/File"));
     Assert.assertNotNull(tFile);
     Assert.assertEquals(getCommandOutput(new String[] {"mkdir",
@@ -348,18 +348,18 @@ public class TFsShellTest {
 
   @Test
   public void mkdirExistingTest() throws IOException {
-    Assert.assertEquals(0, mFsShell.mkdir(new String[] {"mkdir", "/testFile1"}));
-    Assert.assertEquals(0, mFsShell.mkdir(new String[] {"mkdir", "/testFile1"}));
+    Assert.assertEquals(0, mFsShell.run(new String[] {"mkdir", "/testFile1"}));
+    Assert.assertEquals(0, mFsShell.run(new String[] {"mkdir", "/testFile1"}));
   }
 
-  @Test(expected = IOException.class)
+  @Test
   public void mkdirInvalidPathTest() throws IOException {
-    mFsShell.mkdir(new String[] {"mkdir", "/test File Invalid Path"});
+      Assert.assertEquals(-1,mFsShell.run(new String[] {"mkdir", "/test File Invalid Path"}));
   }
 
   @Test
   public void mkdirShortPathTest() throws IOException {
-    mFsShell.mkdir(new String[] {"mkdir", "/root/testFile1"});
+    mFsShell.run(new String[] {"mkdir", "/root/testFile1"});
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/root/testFile1"));
     Assert.assertNotNull(tFile);
     Assert.assertEquals(getCommandOutput(new String[] {"mkdir", "/root/testFile1"}),
@@ -372,7 +372,7 @@ public class TFsShellTest {
     String qualifiedPath =
         "tachyon://" + NetworkUtils.getLocalHostName() + ":" + mLocalTachyonCluster.getMasterPort()
             + "/root/testFile1";
-    mFsShell.mkdir(new String[] {"mkdir", qualifiedPath});
+    mFsShell.run(new String[] {"mkdir", qualifiedPath});
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/root/testFile1"));
     Assert.assertNotNull(tFile);
     Assert
@@ -390,9 +390,9 @@ public class TFsShellTest {
   @Test
   public void renameParentDirectoryTest() throws IOException {
     StringBuilder toCompare = new StringBuilder();
-    mFsShell.mkdir(new String[] {"mkdir", "/test/File1"});
+    mFsShell.run(new String[] {"mkdir", "/test/File1"});
     toCompare.append(getCommandOutput(new String[] {"mkdir", "/test/File1"}));
-    mFsShell.rename(new String[] {"rename", "/test", "/test2"});
+    mFsShell.run(new String[] {"rename", "/test", "/test2"});
     toCompare.append(getCommandOutput(new String[] {"mv", "/test", "/test2"}));
     Assert.assertNotNull(mTfs.getFile(new TachyonURI("/test2/File1")));
     Assert.assertNull(mTfs.getFile(new TachyonURI("/test")));
@@ -403,10 +403,10 @@ public class TFsShellTest {
   @Test
   public void renameTest() throws IOException {
     StringBuilder toCompare = new StringBuilder();
-    mFsShell.mkdir(new String[] {"mkdir", "/testFolder1"});
+    mFsShell.run(new String[] {"mkdir", "/testFolder1"});
     toCompare.append(getCommandOutput(new String[] {"mkdir", "/testFolder1"}));
     Assert.assertNotNull(mTfs.getFile(new TachyonURI("/testFolder1")));
-    mFsShell.rename(new String[] {"rename", "/testFolder1", "/testFolder"});
+    mFsShell.run(new String[] {"rename", "/testFolder1", "/testFolder"});
     toCompare.append(getCommandOutput(new String[] {"mv", "/testFolder1", "/testFolder"}));
     Assert.assertEquals(toCompare.toString(), mOutput.toString());
     Assert.assertNotNull(mTfs.getFile(new TachyonURI("/testFolder")));
@@ -416,23 +416,23 @@ public class TFsShellTest {
   @Test
   public void renameToExistingFileTest() throws IOException {
     StringBuilder toCompare = new StringBuilder();
-    mFsShell.mkdir(new String[] {"mkdir", "/testFolder"});
+    mFsShell.run(new String[] {"mkdir", "/testFolder"});
     toCompare.append(getCommandOutput(new String[] {"mkdir", "/testFolder"}));
-    mFsShell.mkdir(new String[] {"mkdir", "/testFolder1"});
+    mFsShell.run(new String[] {"mkdir", "/testFolder1"});
     toCompare.append(getCommandOutput(new String[] {"mkdir", "/testFolder1"}));
     Assert
-        .assertEquals(-1, mFsShell.rename(new String[] {"rename", "/testFolder1", "/testFolder"}));
+        .assertEquals(-1, mFsShell.run(new String[] {"rename", "/testFolder1", "/testFolder"}));
   }
 
   @Test
   public void rmNotExistingFileTest() throws IOException {
-    Assert.assertEquals(0, mFsShell.rm(new String[] {"rm", "/testFile"}));
+    Assert.assertEquals(0, mFsShell.run(new String[] {"rm", "/testFile"}));
   }
 
   @Test
   public void rmTest() throws IOException {
     StringBuilder toCompare = new StringBuilder();
-    mFsShell.mkdir(new String[] {"mkdir", "/testFolder1/testFolder2/testFile2"});
+    mFsShell.run(new String[] {"mkdir", "/testFolder1/testFolder2/testFile2"});
     toCompare
         .append(getCommandOutput(new String[] {"mkdir", "/testFolder1/testFolder2/testFile2"}));
     TachyonURI testFolder1 = new TachyonURI("/testFolder1");
@@ -441,13 +441,13 @@ public class TFsShellTest {
     Assert.assertNotNull(mTfs.getFile(testFolder1));
     Assert.assertNotNull(mTfs.getFile(testFolder2));
     Assert.assertNotNull(mTfs.getFile(testFile2));
-    mFsShell.rm(new String[] {"rm", "/testFolder1/testFolder2/testFile2"});
+    mFsShell.run(new String[] {"rm", "/testFolder1/testFolder2/testFile2"});
     toCompare.append(getCommandOutput(new String[] {"rm", "/testFolder1/testFolder2/testFile2"}));
     Assert.assertEquals(toCompare.toString(), mOutput.toString());
     Assert.assertNotNull(mTfs.getFile(testFolder1));
     Assert.assertNotNull(mTfs.getFile(testFolder2));
     Assert.assertNull(mTfs.getFile(testFile2));
-    mFsShell.rm(new String[] {"rm", "/testFolder1"});
+    mFsShell.run(new String[] {"rm", "/testFolder1"});
     toCompare.append(getCommandOutput(new String[] {"rm", "/testFolder1"}));
     Assert.assertEquals(toCompare.toString(), mOutput.toString());
     Assert.assertNull(mTfs.getFile(testFolder1));
@@ -458,21 +458,21 @@ public class TFsShellTest {
   @Test
   public void tailLargeFileTest() throws IOException {
     TestUtils.createByteFile(mTfs, "/testFile", WriteType.MUST_CACHE, 2048);
-    mFsShell.tail(new String[] {"tail", "/testFile"});
+    mFsShell.run(new String[] {"tail", "/testFile"});
     byte expect[] = TestUtils.getIncreasingByteArray(1024, 1024);
     Assert.assertArrayEquals(expect, mOutput.toByteArray());
   }
 
   @Test
   public void tailNotExistTest() throws IOException {
-    int ret = mFsShell.tail(new String[] {"tail", "/testFile"});
+    int ret = mFsShell.run(new String[] {"tail", "/testFile"});
     Assert.assertEquals(-1, ret);
   }
 
   @Test
   public void tailSmallFileTest() throws IOException {
     TestUtils.createByteFile(mTfs, "/testFile", WriteType.MUST_CACHE, 10);
-    mFsShell.tail(new String[] {"tail", "/testFile"});
+    mFsShell.run(new String[] {"tail", "/testFile"});
     byte expect[] = TestUtils.getIncreasingByteArray(10);
     Assert.assertArrayEquals(expect, mOutput.toByteArray());
   }
@@ -480,7 +480,7 @@ public class TFsShellTest {
   @Test
   public void touchTest() throws IOException {
     String[] argv = new String[] {"touch", "/testFile"};
-    mFsShell.touch(argv);
+    mFsShell.run(argv);
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/testFile"));
     Assert.assertNotNull(tFile);
     Assert.assertEquals(getCommandOutput(argv), mOutput.toString());
@@ -494,7 +494,7 @@ public class TFsShellTest {
             + mLocalTachyonCluster.getMasterPort() + "/destFileURI";
     // when
     String[] argv = new String[] {"touch", tachyonURI};
-    mFsShell.touch(argv);
+    mFsShell.run(argv);
     // then
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/destFileURI"));
     assertThat(tFile, notNullValue());
